coretelecoms/src/ingestion/requirements.txt
pandas==2.2.2
pyarrow==15.0.0
pyspark==3.5.1
boto3==1.34.0
psycopg2-binary==2.9.9
google-auth==2.29.0
gspread==6.1.2
PyYAML==6.0.1
google-cloud-storage==2.18.2